Rework `QueryProcessor` from a utility class into the interface with several implementations to make the flow clearer.
